+++ src/dreamstate/test-utils/internals.ts
@@ -1,0 +1,4 @@
+/**
+ * Meta symbols for private internals in context managers for test-utils.
+ */
+export const IS_MOCKED: unique symbol = Symbol("IS_MOCKED");