+++ src/dreamstate/test-utils/services/mockManagers.test.ts
@@ -1,14 +1,22 @@
 import { IScopeContext } from "@/dreamstate/core/scoping/ScopeContext";
 import { getCurrent } from "@/dreamstate/test-utils/registry/getCurrent";
-import { mockManagersInScope } from "@/dreamstate/test-utils/registry/mockManagersInScope";
+import { mockScope } from "@/dreamstate/test-utils/registry/mockScope";
+import { mockManagers } from "@/dreamstate/test-utils/services/mockManagers";
+import { TServiceInstanceMap } from "@/dreamstate/types";
 import { ExtendingManager, TestManager } from "@/fixtures";
 
-describe("mockManagersWithScope test util", () => {
+describe("mockManagers test util", () => {
   it("Should properly mock manager with context", () => {
-    const scope: IScopeContext = mockManagersInScope([ TestManager ]);
+    const scope: IScopeContext = mockScope();
+    const map: TServiceInstanceMap = mockManagers([ TestManager ], null, scope);
+
+    expect(map.get(TestManager)).toBeInstanceOf(TestManager);
+    expect(map.get(ExtendingManager)).toBeUndefined();
 
     expect(getCurrent(TestManager, scope)).toBeInstanceOf(TestManager);
     expect(getCurrent(ExtendingManager, scope)).toBeNull();
+
+    expect(getCurrent(TestManager, scope)).toBe(map.get(TestManager));
 
     expect(scope).toBeInstanceOf(Object);
     expect(scope.INTERNAL).toBeInstanceOf(Object);